move extension types
