+++ packages/solana/lib/src/programs/token_2022_program/instruction.dart
@@ -1,5 +1,6 @@
 import 'package:solana/encoder.dart';
 import 'package:solana/solana.dart';
+import 'package:solana/src/programs/token_2022_program/extension_type.dart';
 import 'package:solana/src/programs/token_2022_program/program.dart';
 
 //TODO: token2022 should also have instructions from TokenProgram
@@ -11,51 +12,9 @@ class Token2022Instruction extends Instruction {
           programId: Token2022Program.id,
         );
 
-  factory Token2022Instruction.initializeImmutableOwner({
-    required Ed25519HDPublicKey account,
-  }) =>
-      Token2022Instruction._(
-        accounts: [
-          AccountMeta.writeable(pubKey: account, isSigner: false),
-        ],
-        data: Token2022Program.initializeImmutableOwnerInstructionIndex,
-      );
-
-  factory Token2022Instruction.amountToUiAmount({
-    required Ed25519HDPublicKey mint,
-    required int amount,
-  }) =>
-      Token2022Instruction._(
-        accounts: [
-          AccountMeta.readonly(pubKey: mint, isSigner: false),
-        ],
-        data: ByteArray.merge(
-          [
-            Token2022Program.amountToUiAmountInstructionIndex,
-            ByteArray.u64(amount),
-          ],
-        ),
-      );
-
-  factory Token2022Instruction.uiAmountToAmount({
-    required Ed25519HDPublicKey mint,
-    required int amount,
-  }) =>
-      Token2022Instruction._(
-        accounts: [
-          AccountMeta.readonly(pubKey: mint, isSigner: false),
-        ],
-        data: ByteArray.merge(
-          [
-            Token2022Program.uiAmountToAmountInstructionIndex,
-            ByteArray.u64(amount),
-          ],
-        ),
-      );
-
   factory Token2022Instruction.initializeMintCloseAuthority({
     required Ed25519HDPublicKey mint,
-    required Ed25519HDPublicKey closeAuthority,
+    required Ed25519HDPublicKey? closeAuthority,
   }) =>
       Token2022Instruction._(
         accounts: [
@@ -64,7 +23,7 @@ class Token2022Instruction extends Instruction {
         data: ByteArray.merge(
           [
             Token2022Program.initializeMintCloseAuthorityInstructionIndex,
-            closeAuthority.toByteArray(),
+            if (closeAuthority != null) closeAuthority.toByteArray(),
           ],
         ),
       );
@@ -137,34 +96,3 @@ class Token2022Instruction extends Instruction {
 final nativeMint2022 = Ed25519HDPublicKey.fromBase58(
   '9pan9bMn5HatX4EJdBwg9VgCa7Uz5HL8N1m5D3NdXejP',
 );
-
-enum ExtensionType {
-  uninitialized(0),
-  transferFeeConfig(1),
-  transferFeeAmount(2),
-  mintCloseAuthority(3),
-  confidentialTransferMint(4),
-  confidentialTransferAccount(5),
-  defaultAccountState(6),
-  immutableOwner(7),
-  memoTransfer(8),
-  nonTransferable(9),
-  interestBearingConfig(10),
-  cpiGuard(11),
-  permanentDelegate(12),
-  nonTransferableAccount(13),
-  transferHook(14),
-  transferHookAccount(15),
-  // confidentialTransferFee(16), // Not implemented yet
-  // confidentialTransferFeeAmount(17), // Not implemented yet
-  metadataPointer(18),
-  tokenMetadata(19),
-  groupPointer(20),
-  tokenGroup(21),
-  groupMemberPointer(22),
-  tokenGroupMember(23);
-
-  const ExtensionType(this.value);
-
-  final int value;
-}
